Split into SpotlightContainerDecorator, SpotlightContainer, and useSpotlightContainer
Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)
